requirements.txt
aiogram
spacy
tortoise-orm
newsapi-python
winloop
toml
textblob